Use uintptr_t to cast from void * to unsigned int. This prevents gcc from generating a warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@5398 689a6050-402a-0410-94f2-e92a70836424

+++ action.c
@@ -806,7 +806,7 @@ int do_action(struct action* a, struct sip_msg* msg)
 			}
 
 			if ((a->elem[3].type==NUMBER_ST)) {
-				expires = (unsigned int)a->elem[3].u.data;
+				expires = (uintptr_t)a->elem[3].u.data;
 			}
 
 			ret = cache_store( &a->elem[0].u.s, &name_s, &val_s, expires);
